1st requirement (MVC)
Basic setup with a tap button that increments the score. Testing this is
unnecessarily complex since it requires generating user events in code.
It makes more sense to separate the business logic to a view model,
which can be tested in isolation. VCs are nicer to be tested with UI
tests.