RxTest 1
RxTest to the rescue! Replacing the first unit test, which was simple,
by a new approach using RxTest. It isn't very useful here, but it's
already easier to read, it doesn't require a timeout to be defined, and
we don't need to handle events that we don't want to occur (like in this
case, .error and .completed). We say what must and what must not occur
through an array (not in the array? then it must not occur)

It's also beautiful how we can check timing with perfect precision. A
tap at the instant 5 must increment the score at the instant 5. No error
margin is added at all!