RxTest 2 - failing
Partially replaced the 2nd unit by RxTest, just to see how the events
are generated, but it fails because... There were no events at all! How?
Remember that the timer was defined with MainScheduler, which uses a
real time unit, and in the unit test we're using virtual time... So the
test ended before anything real could be done.