Scheduler injection
If the scheduler was the problem, then easy fix it is! Just basic DI so
that our ViewModel can now use an injected TestScheduler instead of
hardcoding MainScheduler. But it still fails! We expected the first
event to happen at time 5 (the tap time), however it's happening at time
6... Looks like something is incrementing 1 to the virtual time, out of
our control.