Don't simulate processing delay
The TestScheduler internally increments 1 to the virtual time for
certain operations that don't happen immediatelly, in order to better
simulate the real world. This turns out to be a pain depending on the
unit test you're writing, because changes in code will change the
expected times... Better to simply disable that behavior, which is
turned on by default. It still fails, though: now all timer events, from
5.0 until 0.0 and .completed, are happening at time 5. Should we have
kept the processing simulation?